refactor: agent identity cleanup — rename ability, migrations extraction, legacy removal
- Add AgentAbilities class with datamachine/rename-agent and datamachine/list-agents
- Add Agents::update_slug() and Agents::get_all() repository methods
- Rewrite AgentsCommand to delegate to abilities, remove legacy fallback
- Extract 644 lines of migration/scaffold code from data-machine.php to inc/migrations.php
- Register AgentAbilities in plugin bootstrap

The agent slug 'chubesextrachill-com' (derived from sanitize_title on
email login) can now be renamed to something meaningful via:
  wp datamachine agents rename chubesextrachill-com chubes-bot
